Only fetching bookings if connected wallet is owner
Refs #30
